Simplify representation of partitions RDD: each element is a cluster (entity -> records), rather than an entity -> record pair.

+++ src/main/scala/com/github/cleanzr/dblink/analysis/package.scala
@@ -79,7 +79,7 @@ package object analysis {
     import spark.implicits._
     clusters.rdd
       .zipWithUniqueId()
-      .flatMap { case (recIds, entityId) => recIds.iterator.map(recId => (recId, entityId)) }
+      .flatMap { case (recIds, entityId) => recIds.iterator.map(recId => (recId, entityId.toInt)) }
       .toDS()
   }
